+++ Sources/HIGThemesContract/HIGTheme.swift
@@ -48,6 +48,7 @@ public protocol HIGTheme: Sendable {
     var reorderableList: any HIGReorderableListTokens { get }
     var calendar: any HIGCalendarTokens { get }
     var dashboardGrid: any HIGDashboardGridTokens { get }
+    var videoPlayer: any HIGVideoPlayerTokens { get }
     var dataTable: any HIGDataTableTokens { get }
     var breadcrumb: any HIGBreadcrumbTokens { get }
     var pageHeader: any HIGPageHeaderTokens { get }